+++ test_runner/src/main/kotlin/ftl/cli/firebase/test/android/models/AndroidModelsListCommand.kt
@@ -1,0 +1,31 @@
+package ftl.cli.firebase.test.android.models
+
+import ftl.android.AndroidCatalog
+import ftl.args.AndroidArgs
+import ftl.config.FtlConstants
+import picocli.CommandLine
+import java.nio.file.Paths
+
+@CommandLine.Command(
+    name = "list",
+    headerHeading = "",
+    synopsisHeading = "%n",
+    descriptionHeading = "%n@|bold,underline Description:|@%n%n",
+    parameterListHeading = "%n@|bold,underline Parameters:|@%n",
+    optionListHeading = "%n@|bold,underline Options:|@%n",
+    header = ["Print current list of devices available to test against"],
+    description = ["Print current list of Android devices available to test against"],
+    usageHelpAutoWidth = true
+)
+class AndroidModelsListCommand : Runnable {
+    override fun run() {
+        val config = AndroidArgs.load(Paths.get(configPath))
+        println(AndroidCatalog.devicesCatalogAsTable(config.project))
+    }
+
+    @CommandLine.Option(names = ["-c", "--config"], description = ["YAML config file path"])
+    var configPath: String = FtlConstants.defaultAndroidConfig
+
+    @CommandLine.Option(names = ["-h", "--help"], usageHelp = true, description = ["Prints this help message"])
+    var usageHelpRequested: Boolean = false
+}